# ARROW-5921: [C++] Fix multiple nullptr related crashes in IPC