@fsaintjacques 2019-07-12
I merged https://github.com/apache/arrow-testing/pull/4/commits, can you update the submodule?