@wesm 2019-07-12
There was a lint failure

```
+/home/travis/build/apache/arrow/run-cmake-format.py --check
The following cmake files need re-formatting:
- /home/travis/build/apache/arrow/python/CMakeLists.txt
Consider running `run-cmake-format.py`
The command "$TRAVIS_BUILD_DIR/ci/travis_lint.sh" exited with 1.
```

This is my fault -- missed this failure in the patch I merged earlier, I fixed it in master