@wesm 2019-07-12
Not sure why Appveyor is not triggering here but I rebased and pushed a branch on my fork to get a build

https://ci.appveyor.com/project/wesm/arrow/builds/25945385